Add AccurateArithmetic submodule
